server/requirements.txt
aniso8601==7.0.0
beautifulsoup4==4.11.1
black==22.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Cors==3.0.10
Flask-GraphQL==2.0.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.22.4
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
promise==2.3
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
